% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_norm.R
\name{add_cdds}
\alias{add_cdds}
\title{Add range of CDD variables to data frame}
\usage{
add_cdds(
  df,
  tempvar_name = "tavg",
  min_cref = 40,
  max_cref = 80,
  temp_step = 1,
  zero_dd_thresh = NULL
)
}
\arguments{
\item{df}{Data frame with a series of temperature readings.}

\item{tempvar_name}{column name of temperature reading variable.}

\item{min_cref}{Minimum cooling reference temperature to calculate}

\item{max_cref}{Maximum cooling reference temperature to calculate}

\item{temp_step}{Interval by which to increment from \code{min_cref} to
\code{max_cref}.}

\item{zero_dd_thresh}{optional value (0 to 1) setting threshold for minimum
fraction of positive degree day values for candidate balance point
temperatures.}
}
\value{
Data frame with series of variables representing cooling degree day
values calculated from range of cooling reference temperatures from
from \code{min_href} to \code{max_href} by \code{temp_step}.
}
\description{
Calculates the a range of cooling degree day (CDD) values
for a outdoor temperature value across a range of cooling reference
temperatures. The values for each cooling reference temperature are
added as new variables to data frame.
}
\examples{
data(nhtemp)
nhtemp_df = data.frame(tavg = nhtemp)
add_cdds(df = nhtemp_df, temp_step = 4, )
}
