% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_norm.R
\name{add_hdds}
\alias{add_hdds}
\title{Add range of HDD variables to data frame}
\usage{
add_hdds(
  df,
  tempvar_name = "tavg",
  min_href = 40,
  max_href = 80,
  temp_step = 1,
  zero_dd_thresh = NULL
)
}
\arguments{
\item{df}{Data frame with a series of temperature readings.}

\item{tempvar_name}{column name of temperature reading variable.}

\item{min_href}{Minimum heating reference temperature to calculate}

\item{max_href}{Maximum heating reference temperature to calculate}

\item{temp_step}{Interval by which to increment from \code{min_href}
to \code{max_href}.}

\item{zero_dd_thresh}{optional value (0 to 1) setting threshold for minimum
fraction of positive degree day values for candidate balance point
temperatures.}
}
\value{
Data frame with series of variables representing heating degree day
values calculated from range of heating reference temperatures from
from \code{min_href} to \code{max_href} by \code{temp_step}.
}
\description{
Calculates the a range of heating degree day (HDD) values
for a outdoor temperature value across a range of heating reference
temperatures. The values for each heating reference temperature are
added as new variables to data frame.
}
\examples{
require(stoolsr)
data(nhtemp)
nhtemp_df = data.frame(tavg = nhtemp)
add_hdds(df = nhtemp_df, temp_step = 4)
}
