% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cpy_excel}
\alias{cpy_excel}
\title{Write dataframe to the clipboard for pasting in Excel}
\usage{
cpy_excel(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{Data frame}

\item{row.names}{Whether to include row names (default=FALSE)}

\item{col.names}{Whether to include column names (default = TRUE)}

\item{...}{Additional arguments to pass to the write.table function}
}
\value{
Tab-separated table
}
\description{
Write dataframe to the clipboard for pasting in Excel
}
\examples{
df <- data.frame(x = rnorm(10), y = sample(letters, 10))
cpy_excel(df)  # next paste in Excel
}
