% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_tools.R
\name{cpy_header}
\alias{cpy_header}
\title{Write a Script Header to Clipboard}
\usage{
cpy_header()
}
\value{
multi-line string write to clipboard
}
\description{
Write a Script Header to Clipboard
}
\examples{
cpy_header()
}
