% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{gen_flag_ids}
\alias{gen_flag_ids}
\title{Add unique ids to runs of TRUE values in binary vector}
\usage{
gen_flag_ids(x, id_prefix = "x", no_id_tag = NA)
}
\arguments{
\item{x}{a binary vector.}

\item{id_prefix}{string.}

\item{no_id_tag}{label for FALSE condition of binary}
}
\value{
A vector with unique ids for each run of TRUE values \code{x}.
}
\description{
Add unique ids to runs of TRUE values in binary vector
}
\examples{
x = rbinom(20, 1, prob = 0.6)
ids = gen_flag_ids(x)
data.frame(flag = x, id = ids)
}
