% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_norm.R
\name{get_cref}
\alias{get_cref}
\title{Estimate cooling reference temp}
\usage{
get_cref(
  x,
  usage_var = "usage",
  cdd_prefix = "cdd",
  coef_screen = FALSE,
  detailed_results = FALSE,
  calc_loss = FALSE,
  prior_mean = NULL,
  prior_sd = NULL
)
}
\arguments{
\item{x}{Data frame with series of variables representing range of CDD
values for series of billing periods.}

\item{usage_var}{Name of usage variable as string.}

\item{cdd_prefix}{something like "CDD" which is used to select
range of variables for fitting models. Assumes series of variables with
names like "CDD40", "CDD41"..."CDD80".}

\item{coef_screen}{exclude candidate models where one or more coefficients
are negative.}

\item{detailed_results}{return summary object for best candidate model.}

\item{calc_loss}{Use built-in function to calculate liklikhood score. The
default is to select the cooling reference temperature resulting in
largest r-squared value.}

\item{prior_mean}{Estimated mean heating reference temperature of the
population.}

\item{prior_sd}{Estimated standard deviation of the cooling reference
temperature in the population.}
}
\value{
Either named character vector with estimated heating and cooling
reference temps and r-squared value for the best linear model or a
model summary object for the best model if detailed_results selected.
}
\description{
Fits series of linear models across a range of
potential cooling reference temperatures. Returns
temperature with the largest r-squared value or liklihood
score based on a custom loss function.
}
\examples{
cat("TODO: Add example with example monthly usage data.")
}
