% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_norm.R
\name{get_hcref}
\alias{get_hcref}
\title{Estimate reference temps with heating and cooling model}
\usage{
get_hcref(
  x,
  usage_var = "usage",
  hdd_prefix = "hdd",
  cdd_prefix = "cdd",
  coef_screen = FALSE,
  detailed_results = FALSE
)
}
\arguments{
\item{x}{Data frame with series of variables representing range of CDD and
HDD values for series of billing periods.}

\item{usage_var}{Name of usage variable as string.}

\item{hdd_prefix}{something like "HDD" which is used to select
range of variables for fitting models. Assumes series of variables with
names like "HDD40", "HDD41"..."HDD80".}

\item{cdd_prefix}{something like "CDD" which is used to select
range of variables for fitting models. Assumes series of variables with
names like "CDD40", "CDD41"..."CDD80".}

\item{coef_screen}{exclude candidate models where one or more coefficients
are negative.}

\item{detailed_results}{return summary object for best candidate model.}
}
\value{
Either named character vector with estimated heating and cooling
reference temps and r-squared value for the best linear model or a
model summary object for the best model if detailed_results selected.
}
\description{
Fits series of linear models across a range of
potential heating and cooling reference temperatures. Returns
temperature with the largest r-squared value or liklihood
score based on a custom loss function.
}
\examples{
cat("TODO: Add example with example monthly usage data.")

}
