% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_norm.R
\name{get_href}
\alias{get_href}
\title{Estimate heating reference temp}
\usage{
get_href(
  x,
  usage_var = "usage",
  hdd_prefix = "hdd",
  coef_screen = FALSE,
  detailed_results = FALSE,
  calc_loss = FALSE,
  prior_mean = NULL,
  prior_sd = NULL
)
}
\arguments{
\item{x}{Data frame with series of variables representing range of
HDD values for series of billing periods.}

\item{usage_var}{Name of usage variable as string.}

\item{hdd_prefix}{something like "HDD" which is used to select
range of variables for fitting models. Assumes series of variables with
names like "HDD40", "HDD41"..."HDD80".}

\item{coef_screen}{exclude candidate models where one or more coefficients
are negative.}

\item{detailed_results}{return summary object for best candidate model.}

\item{calc_loss}{Use built-in function to calculate liklikhood score. The
default is to select the reference temp resulting in largest r-squared
value.}

\item{prior_mean}{Estimated mean reference temperature of population.}

\item{prior_sd}{Estimated standard deviation of the reference temperature
in the population.}
}
\value{
Either named character vector with estimated heating and cooling
reference temps and r-squared value for the best linear model or a
model summary object for the best model if detailed_results selected.
}
\description{
Fits series of linear models across a range of
potential heating  reference temperatures. Returns
temperature with the largest r-squared value or liklihood
score based on a custom loss function.
}
\examples{
"TODO: Add example with example monthly usage data."
}
