% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_isdlite}
\alias{get_isdlite}
\title{Download and Process Hourly ISDLite Data for Range of Years and Stations}
\usage{
get_isdlite(yr_start, yr_end, stn_ids = NULL, zip5 = NULL, temp_unit = "F")
}
\arguments{
\item{yr_start}{Earliest year for which to pull data}

\item{yr_end}{Most recent year for which to pull data}

\item{stn_ids}{Character combination of usaf and wban code in format "usaf-wban"}

\item{zip5}{5-digit zip code if not including stn_id (can't include both)}

\item{temp_unit}{"F" for degrees Fahrenheit or "C" for Celsius}
}
\value{
data frame
}
\description{
Download and Process Hourly ISDLite Data for Range of Years and Stations
}
\examples{
# Use station_id:
stn_data <- stoolsr::isd_stns()
stns <- stn_data[stn_data$end>=as.Date("2020-01-01"),]
use_stns <- sample(stns$stn_id, 2)
weather_data <- get_isdlite(yr_start = 2019, yr_end = 2020, stn_ids = use_stns)

# Or use zip codes directly which will automatically pull weather from
# nearest station to zip code centroid
weather_data <- get_isdlite(yr_start = 2019, yr_end = 2020, 
 zip5 = c("53715", "53235", "54255"))
}
