% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_tmy}
\alias{get_tmy}
\title{Get TMY Data for Lat/Lon from NREL NSRDB API}
\usage{
get_tmy(lat, lon, api_key = NULL)
}
\arguments{
\item{lat}{numeric latitude for site}

\item{lon}{numeric longitude for site}

\item{api_key}{NREL API key (see:  https://nsrdb.nrel.gov/about/tmy.html)}
}
\value{
list containing 1) TMY data as data frame and 2) metadata as named list
}
\description{
Get TMY Data for Lat/Lon from NREL NSRDB API
}
\examples{
stn_data <- isd_stns()
stns <- stn_data[stn_data$end>=as.Date("2020-01-01"),]
stn <- stns[sample(nrow(stns),1),]

"tmy <- get_tmy(stn$lat, stn$lon, api_key = mykey)"
}
