% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guidehouse_colors}
\alias{guidehouse_colors}
\title{Guidehouse Color Palette}
\format{
A named list:
\describe{
\item{name}{color name}
\item{value}{hexadecimal code}
...
}
}
\usage{
guidehouse_colors
}
\description{
A list of Guidehouse hexadecimal colors
}
\keyword{datasets}
