% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illume_colors}
\alias{illume_colors}
\title{Cadeo Color Palette}
\format{
A named list:
\describe{
\item{name}{color name}
\item{value}{hexadecimal code}
...
}
ILLUME Color Palette

A list of ILLUME hexadecimal colors

A named list:
\describe{
\item{name}{color name}
\item{value}{hexadecimal code}
...
}
}
\usage{
illume_colors
}
\description{
A list of Cadeo hexadecimal colors
}
\keyword{datasets}
