% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_validemail}
\alias{is_validemail}
\title{Valid Email Test}
\usage{
is_validemail(x)
}
\arguments{
\item{x}{email string}
}
\value{
boolean
}
\description{
Valid Email Test
}
\examples{
is_validemail("info@slipstreaminc.org")
is_validemail(c("info@slipstreaminc.org", "info@slipstreaminc.org"))
}
