% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{isd_stns}
\alias{isd_stns}
\title{Load Current ISD Station Meta-data}
\format{
data frame with 12 columns:
\describe{
\item{stns_id}{combined usaf-wban codes string}
\item{usaf}{USAF station code}
\item{wban}{WBAN station code}
\item{staion_name}{NOAA ISD station name}
\item{ctry}{country abbreviation}
\item{state}{state abbreviation}
\item{icao}{ICAO station code}
\item{lat}{station latitude}
\item{lon}{station longitude}
\item{evel_m}{station elevation in meters}
\item{begin}{earliest data with available data}
\item{end}{latest data with available data}
...
}
}
\usage{
isd_stns()
}
\value{
data frame

Source: https://www1.ncdc.noaa.gov/pub/data/noaa/isd-history.csv
}
\description{
Load Current ISD Station Meta-data
}
\examples{
stn_data <- isd_stns()
}
