% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_file_list}
\alias{make_file_list}
\title{Make File List}
\usage{
make_file_list(dir, pattern = NULL)
}
\arguments{
\item{dir}{valid directory as string}

\item{pattern}{regex pattern string to return only file names with string match}
}
\value{
Named list where name is the file name and value is the full file path
}
\description{
Helper function for
}
\examples{
# Ex. List files in current working dir
make_file_list(".")

# Ex. 2:  Return a list of some dummy .csv files with data
dfs <- lapply(rep(50, 20), function(x) data.frame(x= sample(LETTERS, x, replace = TRUE), y = rnorm(x)))
dp <- tempdir()
fps <- file.path(dp, paste0("file", 1:20, ".csv"))
for(i in 1:20) write.csv(dfs[[i]], fps[[i]])
make_file_list(dp, pattern = ".csv")

}
