% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_id}
\alias{make_id}
\title{Create a single alpha-numeric ID}
\usage{
make_id(len = 4, inc_letters = TRUE)
}
\arguments{
\item{len}{Number of characters to include in the ID.}

\item{inc_letters}{Whether to include lower-case letters.}
}
\value{
A string
}
\description{
Create a single alpha-numeric ID
}
\examples{
make_id()
make_id(len = 8, inc_letters = FALSE)

}
