% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthly_gas}
\alias{monthly_gas}
\title{Anonymized of Sample Monthly Gas Consumption Data for MA}
\format{
data frame with 6 columns:
\describe{
\item{dummy_id}{eleven digit pretend account number}
\item{start_date}{start of billing period}
\item{end_date}{end of billing period}
\item{zip5}{five digit zip code}
\item{usage}{therms}
\item{icao}{ICAO station code}
\item{bp_days}{days in billing period}
...
}
}
\usage{
monthly_gas
}
\description{
Data frame with 200 accounts across approx. 2 years (2018-2020)
}
\keyword{datasets}
