% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{read_isdlite}
\alias{read_isdlite}
\title{Read ISD-Lite Hourly Data from NOAA FTP Site}
\usage{
read_isdlite(ftp_url)
}
\arguments{
\item{ftp_url}{ftp path to ".gz" file as character}
}
\value{
data frame
}
\description{
Read ISD-Lite Hourly Data from NOAA FTP Site
}
\examples{
df <- read_isdlite("https://www.ncei.noaa.gov/pub/data/noaa/isd-lite/2021/025420-99999-2021.gz")
}
