% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{robust_se}
\alias{robust_se}
\title{Robust Standard Error}
\usage{
robust_se(model, cluster)
}
\arguments{
\item{model}{model object}

\item{cluster}{grouping variable such as account number or premise}
}
\value{
list of tables
}
\description{
Robust Standard Error
}
\examples{
"TODO: add example"
}
