% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slipstream_colors}
\alias{slipstream_colors}
\title{Slipstream Color Palette}
\format{
A named list:
\describe{
\item{name}{color name}
\item{value}{hexadecimal code}
...
}
}
\usage{
slipstream_colors
}
\description{
A list of Slipstream hexadecimal colors
}
\keyword{datasets}
