% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{vmake_id}
\alias{vmake_id}
\title{Creates a vector of alpha-numeric IDs}
\usage{
vmake_id(size = 10, unique_only = TRUE, len = 4, inc_letters = TRUE)
}
\arguments{
\item{size}{Number of IDs to create.}

\item{unique_only}{Whether to allow duplicate IDs.}

\item{len}{Number of characters to include in the ID.}

\item{inc_letters}{Whether to include lower-case letters.}
}
\value{
Character vector
}
\description{
Creates a vector of alpha-numeric IDs
}
\examples{
vmake_id()
set.seed(1984)
unique_ids <- vmake_id(size = 1e4, unique_only = TRUE, len = 4)
sum(duplicated(unique_ids))
non_unique_ids <-vmake_id(size = 1e4, unique_only = FALSE, len = 4)
sum(duplicated(non_unique_ids))
}
