% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_norm.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize vector}
\usage{
winsorize(x, lower = 0.025, upper = 0.975)
}
\arguments{
\item{x}{Numeric vector.}

\item{lower}{lower quantile}

\item{upper}{upper quantile}
}
\value{
Vector with extreme values set to quantile value. Missing
values also removed.
}
\description{
Sets extreme values to upper and lower quantile.
}
\examples{
 data(precip)
par(mfrow = c(2,1))
hist(precip)
hist(winsorize(precip))
}
